Split functions into separate files.
